+++ example.py
@@ -2,10 +2,10 @@ from banglaspeech2text import available_models, Model
 
 
 models = available_models(True)
-print(models)
-# file_path = r"F:\Code\Python\rodela_bot\app\data\উদ্দেশ্য_licence.wav"
+models['large'][0].show_config()
+file_path = r"F:\Code\Python\rodela_bot\app\data\উদ্দেশ্য_licence.wav"
 
-# model = Model(models[0], verbose=True)
-# model.load()
+model = Model(models[0], verbose=True)
+model.load()
 
-# print(model(file_path))
+print(model(file_path))